+++ RELEASE_NOTES.md
@@ -1,4 +1,25 @@
-### New in 0.9 (not released yet)
+### New in 0.10 (not released yet)
+
+ * Breaking: Updated NuGet reference `Newtonsoft.Json` to v7.0.1
+   (up from v6.0.8)
+ * Breaking: Remove the empty constructor from `SingleValueObject<>`
+ * New: Added `SingleValueObjectConverter` to help create clean JSON when
+   e.g. domain events are serialized
+ * New: Added a protected method `Register(IEventApplier)` to
+   `AggregateRoot<,>` that enables developers to override how events are
+   applied. Use this to e.g. implement state objects
+ * New: Create `AggregateState<,,>` that developers can use to create aggregate
+   state objects. Call `Register(...)` with the state object as argument
+   to redirect events to it
+ * New: Allow `AggregateRoot<,>.Apply(...)`, i.e., methods for applying events,
+   to be `private` and `protected`
+ * New: Made `AggregateRoot<,>.Emit(...)` protected and virtual to allow
+   overrides that e.g. add a standard set of metadata from the aggregate state.
+ * New: Made `AggregateRoot<,>.ApplyEvent(...)` protected and virtual to
+   allow more custom implementations of applying events to the aggregate root.
+ * Fixed: Updated internal NuGet reference `Dapper` to v1.42 (up from v1.38)
+
+### New in 0.9.580 (released 2015-07-20)
 
  * Braking: `IEventStore.LoadAllEventsAsync` and `IEventStore.LoadAllEvents`
    now take a `GlobalPosition` as an argument instead of a `long` for the